Added a full screen card view using the android navigation component
added safeargs to the gradle dependencies
changed the package path name in the manifest
When a card is clicked it now parcels the card data and puts sends it via an action.
Card model can now be parcelled.
Added a full screen card view.
added a main screen host activity.